Move String.splitAt in r2-navigator
