+++ r2-navigator/src/main/java/org/readium/r2/navigator/media/extensions/MediaMetadataCompat.kt
@@ -7,7 +7,7 @@
 package org.readium.r2.navigator.media.extensions
 
 import android.support.v4.media.MediaMetadataCompat
-import org.readium.r2.shared.extensions.splitAt
+import org.readium.r2.navigator.extensions.splitAt
 
 internal val MediaMetadataCompat.id: String? get() =
     getString(MediaMetadataCompat.METADATA_KEY_MEDIA_ID)
